refactor: auth 관련 코드 리팩토링 - 미들웨어에서 불필요한 검증 단계 삭제 - 로그아웃 기능 구현 - / (root path)를 로그인 페이지로 변경하고, 해당 path에 접근하는 어드민 계정을 /main으로 리다이렉팅하도록 수정 - 로그인 플로우에서 ADMIN 계정인지 여부를 확인하도록 개선

+++ src/actions/department.ts
@@ -71,7 +71,7 @@ export async function deleteDepartment(departmentId: string) {
   // Return early if not admin
   const userSession = await getSession();
   // TODO: Decrypt token here and check expiry & admin role
-  if (!userSession) return { isValid: false };
+  if (!userSession.isLoggedIn || !userSession.token) return { isValid: false };
 
   try {
     const res = await api.delete(`/api/v1/department/${departmentId}`, userSession.token);
